fixed app crashing bug
